+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to
     other racks located in the same rows as the infrastructure are represented,
     even if they do not contain infrastructure equipment. Other racks are
     represented when set to true. It is false by default (#61).
+  - Add support for optional alpha channel in `~hexcolor` defined type (#65).
 - pkg:
   - Add dependency on PyGObject python library (used to call Pango library).
   - Add optional dependency on requests-toolbelt python library on web variant